Create upstart scripts that manage running moln and redis.
Move db directory to data.

Log files should have the .log extension, and when parsing the log date
ignore the stdout.log file as well.

Add dev script that will start the foreman process as well.

Make deploy script that will deploy the running application to a set of
servers.