Make requests use a Redis connection pool.
Expose the server configuration as a global, this is used in database to
create pool connections.

Split Config.MaxTimeout into Config.{DB,Server}MaxTimeout since it's
useful to make the server timeout a little longer than the db timeout.

Add Config.DBMaxIdle to control the number of max idle pool connections
to the db.